try to save memory

+++ fairmotion/tasks/motion_prediction/test.py
@@ -172,7 +172,7 @@ def save_motion_files(seqs_T: List[np.ndarray], save_output_path: Path):
 def calculate_metrics(
     pred_seqs: np.ndarray,
     tgt_seqs: np.ndarray
-) -> Dict[int, np.float64]:
+) -> Dict[int, np.float32]:
 
     metric_frames = [6, 12, 18, 24]
     R_pred, _ = conversions.T2Rp(pred_seqs)
@@ -202,7 +202,7 @@ def test_model(
     mean: float,
     std: float,
     max_len: int=None
-) -> Tuple[List[np.ndarray], Dict[int, np.float64]]:
+) -> Tuple[List[np.ndarray], Dict[int, np.float32]]:
     pred_seqs, src_seqs, tgt_seqs = run_model(
         model, dataset, max_len, device, mean, std,
     )
